chore: Refactor in Modules and use Iterators
