+++ src/constants.rs
@@ -1,0 +1,4 @@
+pub static G: f64 = 6.674e-11;
+pub static MASS_EARTH: f64 = 5.972e24;
+pub static MASS_MOON: f64 = 7.34e22;
+pub static MASS_SUN: f64 = 1.989e30;
